+++ lindera-core/src/token_filter.rs
@@ -1,3 +1,5 @@
+use std::ops::Deref;
+
 use crate::{token::Token, LinderaResult};
 
 pub trait TokenFilter: 'static + Send + Sync + TokenFilterClone {
@@ -5,12 +7,28 @@ pub trait TokenFilter: 'static + Send + Sync + TokenFilterClone {
     fn apply<'a>(&self, tokens: &mut Vec<Token<'a>>) -> LinderaResult<()>;
 }
 
+pub struct BoxTokenFilter(Box<dyn TokenFilter + 'static + Send + Sync>);
+
+impl Deref for BoxTokenFilter {
+    type Target = dyn TokenFilter;
+
+    fn deref(&self) -> &dyn TokenFilter {
+        &*self.0
+    }
+}
+
+impl<T: TokenFilter> From<T> for BoxTokenFilter {
+    fn from(token_filter: T) -> BoxTokenFilter {
+        BoxTokenFilter(Box::new(token_filter))
+    }
+}
+
 pub trait TokenFilterClone {
-    fn box_clone(&self) -> Box<dyn TokenFilter + 'static + Send + Sync>;
+    fn box_clone(&self) -> BoxTokenFilter;
 }
 
 impl<T: TokenFilter + Clone + 'static> TokenFilterClone for T {
-    fn box_clone(&self) -> Box<dyn TokenFilter + 'static + Send + Sync> {
-        Box::new(self.clone())
+    fn box_clone(&self) -> BoxTokenFilter {
+        BoxTokenFilter::from(self.clone())
     }
 }