+++ CHANGES.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- Add structure for boxing character filter and token filter #287 @mosuka
+
 ## 0.19.3 (2022-12-26)
 - Bump up version to 0.19.3 #286 @mosuka
 - Refactoring #285 @mosuka